+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Changelog ##
 #### dev ####
+* Add flag for skipping launcher UI
 
 #### v3.13.1 ####
 * Fix loading workshop beta mods when local beta mod should be loaded